feat: check_if_table_has_records; refactor delete_question

+++ data_manager.py
@@ -30,6 +30,12 @@ def get_time():
     return norm_time
 
 
+@connection.connection_handler
+def check_if_table_has_records(cursor, table_name):
+    cursor.execute(f"SELECT EXISTS (SELECT 1 FROM {table_name})")
+    return cursor.fetchone()
+
+
 # ----------------------------------------------------------------------------------------------------
 # Get all question/answer/comment/tag from corresponding tables
 
@@ -142,18 +148,11 @@ def get_next_comment_id(cursor):
 def delete_question(cursor, question_id):
     cursor.execute(f"SELECT image FROM answer WHERE question_id = {question_id}")
     answer_images = cursor.fetchall()
-    cursor.execute(f"""
-    DELETE FROM comment WHERE answer_id IN
-    (SELECT answer.id FROM answer JOIN question ON answer.question_id = question.id
-     WHERE answer.question_id = {question_id})
-    """)
-    cursor.execute(f"DELETE FROM comment WHERE question_id = {question_id}")
-    cursor.execute(f"DELETE FROM question_tag WHERE question_id = {question_id}")
-    cursor.execute(f"DELETE FROM answer WHERE question_id = {question_id}")
     cursor.execute(f"DELETE FROM question WHERE id = {question_id}")
 
     for d in answer_images:
-        os.remove(d['image'])
+        if d['image']:
+            os.remove(d['image'])
 
 
 @connection.connection_handler
